Add special case comparisons of Decimal with NaN/Inf
`Inf` is a common starting value for `min` comparisons (and similarly
for `-Inf` and max). As an example Plots.jl uses it to figure out
the extrema of values to be plotted. Since Inf is not representable
in Decimal, these comparisons were throwing an error. Fix this by
adding special case comparisons with AbstractFloat that behave properly
for the Inf case.